Create separate Tabs component
